requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.6.0
appdirs==1.4.4
appnope==0.1.4
APScheduler==3.10.4
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
asyncio==3.4.3
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.1
cycler==0.12.1
Cython==0.29.36
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
docopt==0.6.2
easydict==1.13
easyocr==1.7.1
editdistance==0.8.1
efficientnet==1.0.0
essential-generators==1.0
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
google-api-core==1.34.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-pubsub==2.1.0
google-pasta==0.2.0
googleapis-common-protos==1.63.2
gotrue==2.8.1
greenlet==3.0.3
grpc-google-iam-v1==0.12.7
grpcio==1.65.1
grpcio-status==1.48.2
gspread==6.1.2
gspread-formatting==1.2.0
h11==0.14.0
h2==4.1.0
h5py==3.10.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.25.2
huggingface-hub==0.24.6
hyperframe==6.0.1
idna==3.10
imageio==2.34.2
imgaug==0.4.0
importlib_metadata==8.2.0
iniconfig==2.0.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==2.15.0
Keras-Applications==1.0.8
keras-ocr==0.9.3
kiwisolver==1.4.5
langdetect==1.0.9
Lasagne @ git+https://github.com/Lasagne/Lasagne.git@61b1ad14d9960a2e20ff2f06ea3f30e81a21441e
lazy_loader==0.4
libclang==18.1.1
libcst==1.4.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.1.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.0
oauth2client==4.1.3
oauthlib==3.2.2
onnx==1.16.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
pkgconfig==1.5.5
platformdirs==4.2.2
playwright==1.47.0
pluggy==1.5.0
postgrest==0.16.11
progressbar2==4.4.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyDrive==1.3.1
pyee==12.0.0
Pygments==2.18.0
pyheif==0.8.0
pyparsing==3.1.2
pypdfium2==4.30.0
pytest==8.3.3
pytest-base-url==2.1.0
pytest-playwright==0.5.2
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-doctr==0.9.0
python-dotenv==1.0.1
python-slugify==8.0.4
python-telegram-bot==20.7
python-utils==3.8.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.9.7
realtime==2.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
shapely==2.0.5
six==1.15.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
sympy==1.13.1
telegram==0.0.1
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.15.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metal==1.1.0
tensorlayerx==0.5.8
termcolor==2.4.0
text-unidecode==1.3
tf2onnx==1.16.1
Theano==0.8.2
threadpoolctl==3.5.0
tifffile==2024.7.2
tinycss2==1.3.0
tomli==2.0.1
torch==2.4.0
torchaudio==0.11.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.19
validators==0.33.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.3
whatsapp-api-client-python==0.0.46
wrapt==1.14.1
yacs==0.1.8
yarg==0.1.9
yarl==1.11.1
zipp==3.19.2
